+++ src/main/java/com/greenfox/lvlup/controller/DTOTestingController.java
@@ -1,6 +1,6 @@
 package com.greenfox.lvlup.controller;
 
-import com.greenfox.lvlup.model.dto.model.ArchetypeDTO;
+import com.greenfox.lvlup.model.dto.ArchetypeDTO;
 import com.greenfox.lvlup.service.ArchetypeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
