+++ src/main/java/com/greenfox/lvlup/model/dto/BadgeLevelDTO.java
@@ -1,0 +1,17 @@
+package com.greenfox.lvlup.model.dto;
+
+public class BadgeLevelDTO {
+
+  private Long id;
+
+  public BadgeLevelDTO() {
+  }
+
+  public Long getId() {
+    return id;
+  }
+
+  public void setId(Long id) {
+    this.id = id;
+  }
+}
