+++ src/main/java/com/greenfox/lvlup/model/entity/Archetype.java
@@ -1,4 +1,6 @@
-package com.greenfox.lvlup.model;
+package com.greenfox.lvlup.model.entity;
+
+import com.greenfox.lvlup.model.entity.BadgeLevel;
 
 import javax.persistence.*;
 import java.util.List;
