+++ src/main/java/com/greenfox/lvlup/model/mockdto/UserBadgeSetDTO.java
@@ -1,4 +1,4 @@
-package com.greenfox.lvlup.model.dto;
+package com.greenfox.lvlup.model.mockdto;
 
 import java.util.ArrayList;
